# openapi.yaml
openapi: 3.0.1
info:
  title: portal NeoHoods API
  version: "2.0"
  description: |
    # Summary
    portal NeoHoods API
servers:
  - url: https://portal.neohoods.com
    description: portal neohoods API
paths:
...
  /hub/custom-pages/{page-ref}:
    parameters:
      - name: page-ref
        in: path
        description: Reference of the custom page
        required: true
        schema:
          type: string
    get:
      operationId: getCustomPage
      summary: Retrieve a custom page by reference
      tags:
        - custom-pages-hub-api
      responses:
        "200":
          description: A single custom page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomPage"
        "404":
          description: Custom page not found
  /admin/custom-pages:
    get:
      operationId: getAdminCustomPages
      summary: Retrieve all custom pages
      tags:
        - custom-pages-admin-api
      responses:
        "200":
          description: A list of custom pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomPage"

    post:
      operationId: createCustomPage
      summary: Create a new custom page
      tags:
        - custom-pages-admin-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomPage"
      responses:
        "201":
          description: Custom page created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomPage"
  /admin/custom-pages/{page-ref}:
    parameters:
      - name: page-ref
        in: path
        description: Reference of the custom page
        required: true
        schema:
          type: string
    get:
      operationId: getAdminCustomPage
      summary: Retrieve a specific custom page by reference
      tags:
        - custom-pages-admin-api
      responses:
        "200":
          description: A single custom page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomPage"
        "404":
          description: Custom page not found

    put:
      operationId: updateCustomPage
      summary: Update a custom page
      tags:
        - custom-pages-admin-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomPage"
      responses:
        "200":
          description: Custom page updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomPage"
        "404":
          description: Custom page not found

    delete:
      operationId: deleteCustomPage
      summary: Delete a custom page
      tags:
        - custom-pages-admin-api
      responses:
        "204":
          description: Custom page deleted successfully
        "404":
          description: Custom page not found
...
  /admin/users/batch:
    post:
      operationId: getUsersByIds
      summary: Retrieve multiple users by their IDs
      tags:
        - users-admin-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userIds:
                  type: array
                  items:
                    type: string
                    format: uuid
              required:
                - userIds
      responses:
        "200":
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/users:
    get:
      operationId: getUsers
      summary: Retrieve all users
      tags:
        - users-admin-api
      responses:
        "200":
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: saveUser
      summary: Save a user
      tags:
        - users-admin-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: User saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{user-id}:
    parameters:
      - name: user-id
        in: path
        description: User ID
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getUser
      summary: Retrieve a specific user by ID
      tags:
        - users-admin-api
      responses:
        "200":
          description: A single user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      summary: Delete a user
      tags:
        - users-admin-api
      responses:
        "204":
          description: User deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/users/{user-id}/password:
    parameters:
      - name: user-id
        in: path
        description: User ID
        required: true
        schema:
          type: string
          format: uuid
    put:
      operationId: setUserPassword
      summary: Set a user's password
      tags:
        - users-admin-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  example: "newSecurePassword"
              required:
                - newPassword
      responses:
        "204":
          description: Password updated successfully
...
components:
...
  securitySchemes:
    BearerAuthOAuth:
      description: Bearer Token.
      type: http
      scheme: bearer
...
  schemas:
...
    User:
      type: object
      required:
        - id
        - username
        - firstName
        - lastName
        - email
        - isEmailVerified
        - disabled
        - type
        - properties
        - streetAddress
        - city
        - postalCode
        - country
      properties:
        id:
          type: string
          format: uuid
        disabled:
          type: boolean
        isEmailVerified:
          type: boolean
        username:
          type: string
        avatarUrl:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        flatNumber:
          type: string
          example: A701
        roles:
          type: array
          items:
            type: string
            example: "admin"
        streetAddress:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
        preferredLanguage:
          type: string
          example: "en"
        profileSharingConsent:
          type: boolean
          description: "User consent for sharing profile information in the community directory"
          default: false
        createdAt:
          type: string
          format: date-time
          description: "User registration date"
        type:
          $ref: "#/components/schemas/UserType"
        properties:
          type: array
          items:
            $ref: "#/components/schemas/Property"
    UserType:
      type: string
      enum:
        - ADMIN
        - OWNER
        - LANDLORD
        - TENANT
        - SYNDIC
        - EXTERNAL
        - CONTRACTOR
        - COMMERCIAL_PROPERTY_OWNER
        - GUEST
      example: "OWNER"
    Property:
      type: object
      required:
        - type
        - name
      properties:
        type:
          $ref: "#/components/schemas/PropertyType"
        name:
          type: string
          example: "Apartment 3A"
    PropertyType:
      type: string
      enum:
        - APARTMENT
        - GARAGE
        - PARKING
        - COMMERCIAL
        - OTHER
      example: "APARTMENT"
...
    CodedError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: A unique error code that identifies the type of error
          example: AUTH001
        message:
          type: string
          description: A human-readable error message
          example: The specified user could not be found
        traceId:
          type: string
          description: A unique identifier for tracing this error instance
          example: 550e8400-e29b-41d4-a716-446655440000
        documentationUrl:
          type: string
          description: URL to the documentation about this error
          example: https://docs.portal.neohoods.com/errors/AUTH001
        variables:
          type: object
          additionalProperties: true
          description: Additional context variables for the error
          example:
            field: "email"
            value: "invalid@email"
...
    CustomPage:
      type: object
      required:
        - id
        - ref
        - position
        - title
        - content
      properties:
        id:
          type: string
          format: uuid
        ref:
          type: string
        order:
          type: integer
        position:
          type: string
          enum: ["footer-links", "copyright", "footer-help"]
        title:
          type: string
        content:
          type: string
...
  responses:
...
    BadRequest:
      description: Bad Request - The request was invalid or cannot be served
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CodedError"
          examples:
            validation_error:
              value:
                code: VAL001
                message: Validation failed
                variables:
                  fields:
                    - field: email
                      message: must be a valid email
            invalid_input:
              value:
                code: VAL002
                message: Missing required field
                variables:
                  field: username
...
    NotFound:
      description: Not Found - The requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CodedError"
          example:
            code: RES001
            message: The requested resource was not found
...